% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_diag_polygenic.r
\name{wos_diag_polygenic}
\alias{wos_diag_polygenic}
\title{'windows of selection' diagram for polygenic}
\usage{
wos_diag_polygenic()
}
\value{
ggplot object
}
\description{
much simpler than the monogenic one
}
\examples{
wos_diag_polygenic()

}
