% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_diagram.r
\name{wos_diagram}
\alias{wos_diagram}
\title{plot idealised diagram to illustrate 'windows of selection'}
\usage{
wos_diagram(conc_n = 100, conc_rr_mort0 = 0.6, conc_ss_mort0 = 0.1,
  mort_slope = 3, sr = TRUE, win_dom_strt = 0.5, rr_cost = 0,
  dom_cost = 0, exposure = 0.5, max_gen = 1000, startfreq = 0.001,
  title = "", addwindow = TRUE, addshading = TRUE,
  addlabels = TRUE, legendpos = "bottom", plot = TRUE, sim = FALSE,
  adv = FALSE, logadv = TRUE, lab_dom = 0)
}
\arguments{
\item{conc_n}{number of x intervals default 50, will influence smoothness of lines, converted to 0-1}

\item{conc_rr_mort0}{concentration killing none of rr}

\item{conc_ss_mort0}{concentration killing none of ss}

\item{mort_slope}{= 3 slope of mortality curves}

\item{sr}{whether to include heterozygotes}

\item{win_dom_strt}{dominance of resistance 0-1 determines position of sr between rr & ss but it's not really dominance}

\item{rr_cost}{cost of resistance to rr (simply added to rr mort)}

\item{dom_cost}{dominance of cost}

\item{exposure}{proportion of popn exposed to insecticide only used in the simulations}

\item{max_gen}{maximum generations to use in the simulations}

\item{startfreq}{starting frequency to use in the simulations}

\item{title}{title for ggplot}

\item{addwindow}{whether to add dotted lines for opening and shutting of window}

\item{addshading}{whether to shade window}

\item{addlabels}{whether to add labels}

\item{legendpos}{where to add legend}

\item{plot}{whether to plot}

\item{sim}{whether to run simulation across the window of selection}

\item{adv}{whether to calculate and output selective advantage}

\item{logadv}{whether to use log scale in selective advantage plots}

\item{lab_dom}{whether to add labels of dominance values, 0 for none, 1 for all, 2 every other}
}
\value{
ggplot object
}
\description{
mortality over time or declining concentration for RR,SR,SS
}
\examples{
wos_diagram(mort_slope=3, win_dom_strt=0.9)
dfsim <- wos_diagram(sim=TRUE, conc_n=10, rr_cost=0) #conc_n=10 num x axis points, makes faster
wos_diagram(adv=TRUE,win_dom_strt=0.1)

}
