% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_plot_input.R
\name{wos_plot_input}
\alias{wos_plot_input}
\title{plot one simulation input for 'windows of selection' following simulation run}
\usage{
wos_plot_input(dfsim, x = "conc", y = "dom_resist", xlab = NULL,
  ylab = NULL, title = "selected simulation inputs", plot = TRUE,
  xreverse = TRUE)
}
\arguments{
\item{dfsim}{data  on concentration and time-to-resistance output from simulation run from wos_sim()}

\item{x}{name of column of x data}

\item{y}{name of column of y data}

\item{xlab}{label of column of x data for plot}

\item{ylab}{label of column of y data for plot}

\item{title}{title for ggplot}

\item{plot}{whether to plot}

\item{xreverse}{whether to reverse x axis for declining concentration}
}
\value{
ggplot object
}
\description{
plot one simulation input for 'windows of selection' following simulation run
}
\examples{
dfsim <- wos_diagram(sim=TRUE, rr_cost=0, plot=FALSE)
wos_plot_input(dfsim, y='dom_resist', ylab='dominance')

}
