% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_plot_sim.r
\name{wos_plot_sim}
\alias{wos_plot_sim}
\title{plot time-to-resistance for 'windows of selection' following simulation run}
\usage{
wos_plot_sim(dfsim, x = "conc", y = "time_to_resistance0.5",
  shape = NULL, ylab = "time to\\nresistance", xlab = NULL,
  xtxt = FALSE, title = "simulation results", plot = TRUE,
  xreverse = TRUE, xlog = FALSE, ylog = FALSE, no_sr = FALSE)
}
\arguments{
\item{dfsim}{data  on concentration and time-to-resistance output from simulation run from wos_sim()}

\item{x}{name of column of x data}

\item{y}{name of column of y data}

\item{shape}{name of column to determine shape of symbols}

\item{ylab}{label of column of y data for plot}

\item{xlab}{label of column of x data for plot}

\item{xtxt}{whether to number x axis}

\item{title}{title for ggplot}

\item{plot}{whether to plot}

\item{xreverse}{whether to reverse x axis for declining concentration}

\item{xlog}{whether to log x axis}

\item{ylog}{whether to log y axis}

\item{no_sr}{option for dealing with data with no SR heteroxzygotes plots dominance 0&1 with diff symbols}
}
\value{
ggplot object
}
\description{
plot time-to-resistance for 'windows of selection' following simulation run
}
\examples{
dfsim <- wos_diagram(sim=TRUE, conc_n=10, rr_cost=0, plot=FALSE)
wos_plot_sim(dfsim)

}
