% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_sim.r
\name{wos_sim}
\alias{wos_sim}
\title{wos_sim run a series of resistance evolution simulations to illustrate 'windows of selection'}
\usage{
wos_sim(dfmortbygen = NULL, x = "concentration", y = "mortality",
  concs = c(1:5), mort_rr = c(0, 0, 0, 0.5, 1), mort_sr = c(0, 0,
  0.5, 1, 1), mort_ss = c(0, 0.5, 1, 1, 1), exposure = 0.5,
  max_gen = 1000, no_thresh = NA, startfreq = 0.001,
  dominances = c(0, 0.1, 1), plot = TRUE)
}
\arguments{
\item{dfmortbygen}{dataframe of mortality by genotype, an alternative way of initialising function, needs genotype column}

\item{x}{column name of x in dfmortbygen}

\item{y}{column name of y in dfmortbygen}

\item{concs}{concentration or time}

\item{mort_rr}{rr mortalities 0-1}

\item{mort_sr}{sr mortalities 0-1}

\item{mort_ss}{ss mortalities 0-1}

\item{exposure}{proportion of popn exposed to insecticide only used in the simulations}

\item{max_gen}{maximum generations to use in the simulations}

\item{no_thresh}{what to return as time-to-resistance if threshold not reached}

\item{startfreq}{starting frequency to use in the simulations}

\item{dominances}{only if there are no SR values which dominances to run}

\item{plot}{whether to plot}
}
\value{
dataframe of simulation outputs
}
\description{
requires either a dataframe with columns for genotype, x (e.g. time) and y (e.g.mortality)
or vectors of x, and yrr,ysr,yss
}
\examples{
dfsim <- wos_sim()
dfsim <- wos_sim(dfmortbygen=dplyr::data_frame(conc=rep(c(1,2),3),
                               genotype=rep(c('RR','SR','SS'),each=2),
                               mort=c(0,0.2,0.4,0.6,0.8,1)),
        x='conc',y='mort')
#with no sr data runs dom0,0.1&1

}
